increment 12: rework in the quest for magic singleton
- this singleton is env package.
I moved fields of the config that were filled in the runtime
to the env; and most importantly, the db handle is now in env too.
Now it was only logical to pass to storage.New the instance
of this env instead of config as before (db handle needed by db storage).
Then I tried to store the storage object in the env too and failed
due to circular imports:
- type Environment had *storage.Storage field, so env imported storage for this type
- storage imported env.Environment type for methods New, NewFileStorage, NewDBStorage

I didn't figure out how to uncycle it, so I still pass around two
objects - env and storage